+++ source/intermediate_representation/code_generation/machine_block.h
@@ -13,24 +13,26 @@ namespace ir::cg {
 
 		void set_live_in(const utility::dense_set& live_in);
 		void set_first(handle<instruction> first);
-		void set_terminator(i32 terminator);
-		void set_start(i32 start);
-		void set_end(i32 end);
+		void set_terminator(u64 terminator);
+
+		void set_start(u64 start);
+		void set_end(u64 end);
 
 		handle<instruction> get_first() const;
 		utility::dense_set& get_live_out();
 		utility::dense_set& get_live_in();
 		utility::dense_set& get_kill();
 		utility::dense_set& get_gen();
-		i32 get_terminator() const;
-		i32 get_start() const;
-		i32 get_end() const;
+		u64 get_terminator() const;
+
+		u64 get_start() const;
+		u64 get_end() const;
 	private:
 		// TODO: replace by a linked list of instructions 
 		handle<instruction> m_first;
-		i32 m_terminator = 0;
-		i32 m_start = 0;
-		i32 m_end = 0;
+		u64 m_terminator = 0;
+		u64 m_start = 0;
+		u64 m_end = 0;
 
 		utility::dense_set m_live_out;
 		utility::dense_set m_live_in;
